chore: merge staging to main

+++ api/core/dependencies/job_runner/tasks/ai_tools/tweet_to_tiktok.py
@@ -9,6 +9,7 @@ from api.v1.services.tools.talking_avatar import talking_avatar_service
 from api.db.database import get_db
 from api.utils.minio_service import minio_service
 from api.utils.settings import settings
+from api.utils.replicate_service import replicate_service
 from api.utils import mime_types
 from api.v1.services.job import tifi_job_service
 from api.core.dependencies.job_runner.app.utils import save_and_print_job_progress
@@ -36,24 +37,20 @@ media_urls = payload.get('scene_media_urls')
 video_style = payload.get('video_style')
 voice_url = payload.get('voice_url')
 avatar_image_url = payload.get('avatar_image_url')
+width = payload.get('width')
+height = payload.get('height')
 # voice_over = payload.get('voice_over')
 
+audio_file = None
+
 try:
-    
-    save_and_print_job_progress(db, job, 15, 'Generating audio from script')
-    audio_file = tweet_to_tiktok_service.generate_audio(script, voice_url)
-    # audio_file = os.path.join('tst_scripts', 'results', 'testing.wav')
-    
-    save_and_print_job_progress(db, job, 20, 'Generating subtitle file from generated audio')
-    subtitles_file = tweet_to_tiktok_service.generate_subtitles(audio_file)
-    
     # if video_style == 'talking avatar':
     if avatar_image_url:
-        save_and_print_job_progress(db, job, 25, f'Downloading and opening avatar image file from {avatar_image_url}')
-        image_file = minio_service.download_file_from_minio(avatar_image_url)
+        save_and_print_job_progress(db, job, 15, 'Converting text to speech')
+        audio_url = replicate_service.convert_text_to_speech(text=script, sample_audio_file=voice_url)
         
         save_and_print_job_progress(db, job, 30, 'Generating talking avatar video')
-        video_url = talking_avatar_service.generate_talking_avatar_video(image_file, audio_file)
+        video_url = replicate_service.generate_talking_avatar(avatar_image_url, audio_url)
 
         save_and_print_job_progress(db, job, 45, 'Downloading and saving generated video')
         # Download video file to the current directory
@@ -62,32 +59,40 @@ try:
             extension='mp4',
             prefix_file_name='video'
         )
+        
     else:
+        save_and_print_job_progress(db, job, 15, 'Generating audio from script')
+        audio_file = tweet_to_tiktok_service.generate_audio(script, voice_url)
+        # audio_file = os.path.join('tst_scripts', 'results', 'testing.wav')
+        
         save_and_print_job_progress(db, job, 25, 'Downloading media files')
         media_files = tweet_to_tiktok_service.download_media(media_urls)
 
         save_and_print_job_progress(db, job, 45, 'Composing video with media files')
         video_file = tweet_to_tiktok_service.compose_video(
             audio_file, media_files,
-            width=int(1080 * 0.75),
-            height=int(1920* 0.75)
+            width=width,
+            height=height
         )
 
-    save_and_print_job_progress(db, job, 60, 'Embedding subtitles in generated video')
+    save_and_print_job_progress(db, job, 60, 'Generating subtitle file from generated audio')
+    subtitles_file = tweet_to_tiktok_service.generate_subtitles(audio_file, width, height)
+    
+    save_and_print_job_progress(db, job, 70, 'Embedding subtitles in generated video')
     video_with_subtitles = tweet_to_tiktok_service.add_subtitles_to_video(subtitles_file, video_file)
 
     if background_audio:
-        save_and_print_job_progress(db, job, 70, 'Applying background music to the video')
+        save_and_print_job_progress(db, job, 80, 'Applying background music to the video')
         # Add background music to video
         video_with_audio = video_service.add_background_audio(
             video_path=video_with_subtitles, 
             audio_path=background_audio, 
         )
 
-    save_and_print_job_progress(db, job, 80, 'Cleaning up')
+    save_and_print_job_progress(db, job, 90, 'Cleaning up')
     # Delete unnecessary files
     # TODO: Uncomment this
-    if 'testing' not in audio_file:
+    if audio_file and 'testing' not in audio_file:
         delete_file(audio_file)
     delete_file(subtitles_file)
     delete_file(video_file)
